+++ ivy/array/experimental/random.py
@@ -159,3 +159,60 @@ class ArrayWithRandomExperimental(abc.ABC):
             seed=seed,
             out=out,
         )
+
+    def poisson(
+            self: ivy.Array,
+            *,
+            shape: Optional[Union[ivy.Shape, ivy.NativeShape]] = None,
+            device: Optional[Union[ivy.Device, ivy.NativeDevice]] = None,
+            dtype: Optional[Union[ivy.Dtype, ivy.NativeDtype]] = None,
+            seed: Optional[int] = None,
+            out: Optional[ivy.Array] = None,
+    ):
+        """
+
+        Parameters
+        ----------
+
+        self
+            Input Array of rate paramter(s). It must have a shape that is broadcastable
+            to the requested shape
+        shape
+            If the given shape is, e.g '(m, n, k)', then 'm * n * k' samples are drawn.
+            (Default value = 'None', where 'ivy.shape(lam)' samples are drawn)
+        device
+            device on which to create the array 'cuda:0', 'cuda:1', 'cpu' etc.
+            (Default value = None).
+        dtype
+            output array data type. If ``dtype`` is ``None``, the output array data
+            type will be the default floating-point data type. Default ``None``
+        seed
+            A python integer. Used to create a random seed distribution
+        out
+            optional output array, for writing the result to. It must have a shape that the
+            inputs broadcast to.
+
+        Returns
+        -------
+        ret
+            Drawn samples from the parameterized poisson distribution.
+
+        Examples
+        --------
+        >>> lam = ivy.array([1.0, 2.0, 3.0])
+        >>> lam.poisson()
+        ivy.array([1., 4., 4.])
+
+        >>> lam = ivy.array([1.0, 2.0, 3.0])
+        >>> lam.poisson(shape=(2,3))
+        ivy.array([[0., 2., 2.],
+                   [1., 2., 3.]])
+        """
+        return ivy.poisson(
+            self,
+            shape=shape,
+            device=device,
+            dtype=dtype,
+            seed=seed,
+            out=out,
+        )
