+++ ivy/functional/backends/jax/experimental/random.py
@@ -8,7 +8,12 @@ import jaxlib.xla_extension
 import ivy
 from ivy.functional.backends.jax import JaxArray
 from ivy.functional.backends.jax.random import RNG, _setRNG, _getRNG  # noqa
-from ivy.functional.ivy.random import _check_bounds_and_get_shape
+from ivy.func_wrapper import with_unsupported_dtypes
+from .. import backend_version
+from ivy.functional.ivy.random import (
+    _check_bounds_and_get_shape,
+    _check_shapes_broadcastable,
+)
 from ivy.functional.backends.jax.device import to_device
 
 # Extra #
@@ -69,3 +74,28 @@ def gamma(
     if seed is not None:
         jax.random.PRNGKey(seed)
     return to_device(jax.random.gamma(rng_input, alpha, beta, shape, dtype), device)
+
+
+@with_unsupported_dtypes({"0.3.14 and below": ("bfloat16",)}, backend_version)
+def poisson(
+        lam: Union[float, JaxArray],
+        *,
+        shape: Optional[Union[ivy.NativeShape, Sequence[int]]] = None,
+        device: jaxlib.xla_extension.Device,
+        dtype: jnp.dtype = None,
+        seed: Optional[int] = None,
+        out: Optional[JaxArray] = None,
+) -> JaxArray:
+    lam = jnp.array(lam)
+    _check_shapes_broadcastable(shape, lam.shape)
+    if seed:
+        rng_input = jax.random.PRNGKey(seed)
+    else:
+        RNG_, rng_input = jax.random.split(_getRNG())
+        _setRNG(RNG_)
+    return (
+        to_device(
+            jax.random.poisson(rng_input, lam, shape=shape),
+            device,
+        )
+    )
