+++ ivy/functional/backends/numpy/experimental/random.py
@@ -4,7 +4,12 @@ import numpy as np
 
 # local
 import ivy
-from ivy.functional.ivy.random import _check_bounds_and_get_shape
+from ivy.func_wrapper import with_unsupported_dtypes
+from .. import backend_version
+from ivy.functional.ivy.random import (
+    _check_bounds_and_get_shape,
+    _check_shapes_broadcastable,
+)
 
 
 # dirichlet
@@ -59,3 +64,20 @@ def gamma(
     if seed is not None:
         np.random.seed(seed)
     return np.asarray(np.random.gamma(alpha, beta, shape), dtype=dtype)
+
+
+@with_unsupported_dtypes({"1.23.0 and below": ("bfloat16",)}, backend_version)
+def poisson(
+        lam: Union[float, np.ndarray],
+        *,
+        shape: Optional[Union[ivy.NativeArray, Sequence[int]]] = None,
+        device: str,
+        dtype: np.dtype,
+        seed: Optional[int] = None,
+        out: Optional[np.ndarray] = None,
+) -> np.ndarray:
+    lam = np.array(lam)
+    _check_shapes_broadcastable(shape, lam.shape)
+    if seed:
+        np.random.seed(seed)
+    return np.asarray(np.random.poisson(lam, shape), dtype=dtype)
