+++ ivy/functional/backends/tensorflow/experimental/random.py
@@ -9,11 +9,14 @@ from tensorflow.python.framework.dtypes import DType
 import ivy
 from .. import backend_version
 from ivy.func_wrapper import with_unsupported_dtypes
-from ivy.functional.ivy.random import _check_bounds_and_get_shape
+from ivy.functional.ivy.random import (
+    _check_bounds_and_get_shape,
+    _check_shapes_broadcastable,
+)
 
 
 # dirichlet
-@with_unsupported_dtypes({"2.9.1 and below": ("blfoat16", "float16")}, backend_version)
+@with_unsupported_dtypes({"2.9.1 and below": ("blfoat16", "float16",)}, backend_version)
 def dirichlet(
     alpha: Union[tf.Tensor, tf.Variable, float, Sequence[float]],
     /,
@@ -83,3 +86,24 @@ def gamma(
     beta = tf.cast(beta, dtype)
     with tf.device(device):
         return tfp.distributions.Gamma(alpha, beta).sample(shape, seed=seed)
+
+
+@with_unsupported_dtypes({"2.9.1 and below": ("bfloat16",)}, backend_version)
+def poisson(
+        lam: Union[float, tf.Tensor, tf.Variable],
+        *,
+        shape: Optional[Union[ivy.NativeShape, Sequence[int]]] = None,
+        device: str,
+        dtype: DType,
+        seed: Optional[int] = None,
+        out: Optional[Union[tf.Tensor, tf.Variable]] = None,
+):
+    lam = tf.cast(lam, "float32")
+    with tf.device(device):
+        if seed:
+            tf.random.set_seed(seed)
+        if shape is None:
+            return tf.random.poisson((), lam, dtype=dtype, seed=seed)
+        _check_shapes_broadcastable(shape, lam.shape)
+        lam = tf.broadcast_to(lam, tuple(shape))
+        return tf.random.poisson((), lam, dtype=dtype, seed=seed)
