+++ ivy/functional/backends/tensorflow/random.py
@@ -162,24 +162,3 @@ def shuffle(
     if seed:
         tf.random.set_seed(seed)
     return tf.random.shuffle(x, seed=seed)
-
-
-@with_unsupported_dtypes({"2.9.1 and below": ("bfloat16",)}, backend_version)
-def poisson(
-    lam: Union[float, tf.Tensor, tf.Variable],
-    *,
-    shape: Optional[Union[ivy.NativeShape, Sequence[int]]] = None,
-    device: str,
-    dtype: DType,
-    seed: Optional[int] = None,
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-):
-    lam = tf.cast(lam, "float32")
-    with tf.device(device):
-        if seed:
-            tf.random.set_seed(seed)
-        if shape is None:
-            return tf.random.poisson((), lam, dtype=dtype, seed=seed)
-        _check_shapes_broadcastable(shape, lam.shape)
-        lam = tf.broadcast_to(lam, tuple(shape))
-        return tf.random.poisson((), lam, dtype=dtype, seed=seed)
