+++ ivy/functional/backends/torch/random.py
@@ -10,7 +10,6 @@ from ivy.functional.ivy.random import (
     _check_bounds_and_get_shape,
     _randint_check_dtype_and_bound,
     _check_valid_scale,
-    _check_shapes_broadcastable,
 )
 from ivy.func_wrapper import with_unsupported_dtypes
 from . import backend_version
@@ -132,23 +131,3 @@ def shuffle(
 
 
 shuffle.support_native_out = True
-
-
-@with_unsupported_dtypes({"1.11.0 and below": ("bfloat16",)}, backend_version)
-def poisson(
-    lam: Union[float, torch.Tensor],
-    *,
-    shape: Optional[Union[ivy.NativeShape, Sequence[int]]] = None,
-    device: torch.device,
-    dtype: torch.dtype,
-    seed: Optional[int] = None,
-    out: Optional[torch.Tensor] = None,
-):
-    lam = torch.tensor(lam, device=device, dtype=torch.float32)
-    if seed:
-        torch.manual_seed(seed)
-    if shape is None:
-        return torch.poisson(lam).type(dtype).to(device)
-    _check_shapes_broadcastable(shape, lam.shape)
-    lam = torch.broadcast_to(lam, tuple(shape))
-    return torch.poisson(lam).type(dtype).to(device)
