+++ ivy/functional/ivy/random.py
@@ -595,95 +595,3 @@ def shuffle(
     }
     """
     return ivy.current_backend(x).shuffle(x, seed=seed, out=out)
-
-
-@to_native_arrays_and_back
-@handle_out_argument
-@infer_device
-@infer_dtype
-@handle_nestable
-@handle_exceptions
-def poisson(
-    lam: Union[float, ivy.Array, ivy.NativeArray],
-    *,
-    shape: Optional[Union[ivy.Shape, ivy.NativeShape]] = None,
-    device: Optional[Union[ivy.Device, ivy.NativeDevice]] = None,
-    dtype: Optional[Union[ivy.Dtype, ivy.NativeDtype]] = None,
-    seed: Optional[int] = None,
-    out: Optional[ivy.Array] = None,
-) -> ivy.Array:
-    """
-    Draws samples from a poisson distribution.
-
-    Parameters
-    ----------
-    lam
-        Rate parameter(s) describing the poisson distribution(s) to sample. It must have 
-        a shape that is broadcastable to the requested shape
-    shape
-        If the given shape is, e.g '(m, n, k)', then 'm * n * k' samples are drawn.
-        (Default value = 'None', where 'ivy.shape(lam)' samples are drawn)
-    device
-        device on which to create the array 'cuda:0', 'cuda:1', 'cpu' etc.
-        (Default value = None).
-    dtype
-        output array data type. If ``dtype`` is ``None``, the output array data
-        type will be the default floating-point data type. Default ``None``
-    seed
-        A python integer. Used to create a random seed distribution
-    out
-        optional output array, for writing the result to. It must have a shape that the
-        inputs broadcast to.
-
-    Returns
-    -------
-     ret
-        Drawn samples from the poisson distribution
-
-    Functional Examples
-    -------------------
-    >>> lam = 2.
-    >>> ivy.poisson(lam)
-    ivy.array(1.)
-
-    >>> lam = ivy.array([[1.,2.], [2.,3.]])
-    >>> ivy.poisson(lam)
-    ivy.array([[1., 1.],
-              [0., 4.]])
-
-    >>> lam = ivy.array([[1.,2.], [2.,3.]])
-    >>> ivy.poisson(lam, shape=(3,2,2))
-    ivy.array([[[1., 5.],
-               [1., 2.]],
-               [[0., 0.],
-               [2., 6.]],
-               [[1., 6.],
-               [2., 4.]]])
-
-    >>> lam = ivy.array([[1.,2.], [2.,3.]])
-    >>> ivy.poisson(lam, shape=(3,2,2), seed=47)
-    ivy.array([[[0., 5.],
-               [2., 1.]],
-               [[0., 2.],
-               [2., 2.]],
-               [[1., 5.],
-               [2., 5.]]])
-
-    >>> lam = ivy.Container(a=ivy.array([[1,2], [4,5], [3,1]]),
-    ...                     b=ivy.array([1,4,3,2]))
-    >>> ivy.poisson(lam, dtype=ivy.float16)
-    {
-        a: ivy.array([[1., 5.],
-                      [6., 9.],
-                      [2., 2.]]),
-        b: ivy.array([1., 3., 4., 4.])
-    }
-    """
-    return ivy.current_backend().poisson(
-        lam,
-        shape=shape,
-        device=device,
-        dtype=dtype,
-        seed=seed,
-        out=out,
-    )
