+++ ivy_tests/test_ivy/test_functional/test_experimental/test_core/test_random.py
@@ -167,3 +167,62 @@ def test_gamma(
     for (u, v) in zip(ret, ret_gt):
         assert ivy.all(u >= 0)
         assert ivy.all(v >= 0)
+
+
+# poisson
+@handle_test(
+    fn_tree="functional.ivy.experimental.poisson",
+    dtype_and_lam=helpers.dtype_and_values(
+        available_dtypes=helpers.get_dtypes("float", full=False),
+        min_value=0,
+        max_value=5,
+        min_num_dims=0,
+    ),
+    shape=helpers.get_shape(
+        min_num_dims=1,
+        max_num_dims=3,
+        min_dim_size=1,
+        max_dim_size=10,
+    ),
+    dtype=helpers.get_dtypes("float", full=False),
+    seed=helpers.ints(min_value=0, max_value=100),
+)
+def test_poisson(
+        *,
+        dtype_and_lam,
+        shape,
+        dtype,
+        seed,
+        test_flags,
+        backend_fw,
+        fn_name,
+        on_device,
+        ground_truth_backend,
+):
+    lam_dtype, lam = dtype_and_lam
+    shape = shape + ivy.shape(lam[0])
+
+    def call():
+        return helpers.test_function(
+            ground_truth_backend=ground_truth_backend,
+            input_dtypes=lam_dtype,
+            test_flags=test_flags,
+            on_device=on_device,
+            fw=backend_fw,
+            fn_name=fn_name,
+            test_values=False,
+            lam=lam[0],
+            shape=shape,
+            dtype=dtype[0],
+            seed=seed,
+        )
+
+    ret, ret_gt = call()
+    if seed:
+        ret1, ret_gt1 = call()
+        assert ivy.any(ret == ret1)
+    ret = helpers.flatten_and_to_np(ret=ret)
+    ret_gt = helpers.flatten_and_to_np(ret=ret_gt)
+    for (u, v) in zip(ret, ret_gt):
+        assert u.dtype == v.dtype
+        assert u.shape == v.shape
